BIG BOY UPDATE
Haven't committed to GitHub in a while so a lot of things changed. 
The first of two large items is the new inventory class, item class, and items.js file. These make up the current inventory system (currently unused in the game) but necessary for further progress on systems such as the market and the building system. 
Secondly, the Building, building template, and building shape classes have been added. the building shape is a class that holds a set of x position and y position offsets on the tile size grid in game. These shapes are used to outline where the build is going to be placed (on the mouse cursor) through use of the ____.drawHighlight() function. building templates are the basic definitions of buildings with info about the shape, size, employees, tasks, costs, etc. that can be assigned to new Building classes that function as the in game equivalents.